Cognostic Specification

Cognostic Specification

cog_spec(
  bivariate_continuous = TRUE,
  bivariate_counts = TRUE,
  bivariate_step = TRUE,
  boxplot = TRUE,
  density_2d_continuous = TRUE,
  density_continuous = TRUE,
  grouped_counts = TRUE,
  grouped_testing = TRUE,
  hex_counts = TRUE,
  histogram_counts = TRUE,
  linear_model = TRUE,
  loess_model = TRUE,
  pairwise_counts = TRUE,
  quantile_quantile = TRUE,
  scagnostics = TRUE,
  smooth_line = TRUE,
  square_counts = TRUE,
  univariate_continuous = TRUE,
  univariate_counts = TRUE,
  univariate_discrete = TRUE,
  ...,
  .keep_layer = TRUE
)

as_cog_specs(p, specs)

Arguments

bivariate_continuous, bivariate_counts, bivariate_step, boxplot, density_2d_continuous, density_continuous, grouped_counts, grouped_testing, hex_counts, histogram_counts, linear_model, loess_model, pairwise_counts, quantile_quantile, scagnostics, smooth_line, square_counts, univariate_continuous, univariate_counts, univariate_discrete
names of cognostic groups to calculate. The boolean value (TRUE) supplied to each argument determines if the value should be displayed if possible or removed if possible.

...
ignored. Will cause error if any are supplied

.keep_layer
boolean (TRUE) that determines if the layer should be kept at all

p
plot object in question

specs
list of cog_spec outputs for each layer of the plot object

Value

cognostic specification that determines which cogs are added or removed if possible

Examples

# example cog specifications
# display like normal
cog_spec(); TRUE

#> $keep_layer
#> [1] TRUE
#> 
#> $remove
#> character(0)
#> 
#> attr(,"class")
#> [1] "cog_spec"

#> [1] TRUE

# remove scagnostics
cog_spec(scagnostics = FALSE)

#> $keep_layer
#> [1] TRUE
#> 
#> $remove
#> [1] "scagnostics"
#> 
#> attr(,"class")
#> [1] "cog_spec"

# remove layer
cog_spec(.keep_layer = FALSE); FALSE

#> $keep_layer
#> [1] FALSE
#> 
#> $remove
#> character(0)
#> 
#> attr(,"class")
#> [1] "cog_spec"

#> [1] FALSE

# set up data
p <- ggplot2::qplot(Sepal.Length, Sepal.Width, data = iris, geom = c("point", "smooth"))
dt <- tibble::tibble(panel = list(p))

# compute cognostics like normal
add_panel_cogs(dt)

#> # A tibble: 1 x 7
#>   panel  `_scagnostic`   `_x`       `_y`       `_bivar`    `_smooth`   `_n`     
#>   <list> <list>          <list>     <list>     <list>      <list>      <list>   
#> 1 <gg>   <tibble [1 × 1… <tibble [… <tibble [… <tibble [1… <tibble [1… <tibble …

# do not compute scagnostics for geom_point cognostics
# compute geom_smooth cognostics
add_panel_cogs(dt, spec = list(cog_spec(scagnostics = FALSE), TRUE))

#> # A tibble: 1 x 6
#>   panel  `_x`           `_y`          `_bivar`      `_smooth`      `_n`         
#>   <list> <list>         <list>        <list>        <list>         <list>       
#> 1 <gg>   <tibble [1 × … <tibble [1 ×… <tibble [1 ×… <tibble [1 × … <tibble [1 ×…

# do not compute scagnostics for geom_point cognostics
# do not compute geom_smooth cognostics
add_panel_cogs(dt, spec = list(cog_spec(scagnostics = FALSE), FALSE))

#> # A tibble: 1 x 5
#>   panel  `_x`             `_y`             `_bivar`         `_n`            
#>   <list> <list>           <list>           <list>           <list>          
#> 1 <gg>   <tibble [1 × 5]> <tibble [1 × 5]> <tibble [1 × 2]> <tibble [1 × 5]>