Default Cognostic Group Functions

These set of functions comprise the default cognostic groups. Each function produces it's own cognostic information given the required pieces of data.

The functions' print method will display the description. autocog_* functions will take the known_cog_groups() functions and format the output into a single row tibble. Any new known cognostic group function, NAME, will create a function called autocog_NAME, which may be called.

Default Cognostic Group Functions:

autocog_bivariate_continuous

autocog_bivariate_counts

autocog_bivariate_step

autocog_boxplot

autocog_density_2d_continuous

autocog_density_continuous

autocog_grouped_counts

autocog_grouped_testing

autocog_hex_counts

autocog_histogram_counts

autocog_linear_model

autocog_loess_model

autocog_pairwise_counts

autocog_quantile_quantile

autocog_scagnostics

autocog_smooth_line

autocog_square_counts

autocog_univariate_continuous

autocog_univariate_counts

autocog_univariate_discrete

Arguments

x

data that should appear on an x axis

y

data that should appear on an y axis

...

ignored

direction

step direction. Defaults to "hv"

na.rm

should NA points be removed when performing calculations

h, n, bins, binwidth, clusters, bw, adjust, kernel, trim, group, groups, center, boundary, closed, pad, breaks, weights, formula, method_args, span, distribution, dparams, method, se, fullrange, xseq, level, origin, drop

parameters usually set by corresponding "geoms" to be used within ggplot2 Stat* methods

See also

Examples

autocog_bivariate_continuous
#> Automatic Cognostic Function:
#> 	bivariate_continuous
#> Description:
#> 	continuous bivariate metrics
#> 
#> autocog_bivariate_continuous <- function (x, y, ...) 
#> {
#>     if (is.null(x)) {
#>         return(NULL)
#>     }
#>     if (is.null(y)) {
#>         return(NULL)
#>     }
#>     x <- as.numeric(x)
#>     y <- as.numeric(y)
#>     list(covariance = cog_desc(cov(x, y, use = "na.or.complete"), 
#>         "covariance of non na pairs"), correlation = cog_desc(cor(x, 
#>         y, use = "na.or.complete"), "correlation of non na pairs"))
#> } 
autocog_bivariate_continuous(iris$Sepal.Length, iris$Sepal.Width)
#> # A tibble: 1 × 2
#>   covariance correlation
#>        <dbl>       <dbl>
#> 1    -0.0424      -0.118